Reorganized project structure: grouped assets and scripts
