+++ scripts/colors.py
