+++ scripts/difficulty_multiplyer.py
