+++ scripts/effect_provider.py
@@ -1,4 +1,4 @@
-from enum import Enum
+from scripts.enum import Enum
 import random
 
 class Effect(Enum):
@@ -6,5 +6,5 @@ class Effect(Enum):
     Toxic = 2
 
 def get_random_effect():
-    # return random value from Effect enumeration
+    # return random value from scripts.Effect enumeration
     return random.choice(list(Effect))
